# api/openapi.yaml
openapi: "3.0.3"
info:
  title: HRMS API
  version: "1.0.0"
  description: Human Resource Management System API
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Exchange a valid refresh token for a new token pair (rotation).
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Tokens rotated successfully
          content:
            application/json:
              schema:
                $ref: "schemas/auth/refresh.yaml#/RefreshTokenResponse"
          headers:
            Set-Cookie:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/CommonError"
        "500":
          $ref: "#/components/responses/CommonError"
...
components:
  responses:
...
    CommonError:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "schemas/common.yaml#/ErrorResponse"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token